fix: moved logic to camuda 7 process adapter
