--- connect/connect-c7/process-adapter-in-c7/src/main/java/io/miragon/miranum/connect/adapter/in/c7/message/Camunda7ProcessAutoConfiguration.java
@@ -1,13 +1,0 @@
-package io.miragon.miranum.connect.adapter.in.c7.message;
-
-import org.springframework.boot.context.properties.EnableConfigurationProperties;
-import org.springframework.context.annotation.Bean;
-
-@EnableConfigurationProperties
-public class Camunda7ProcessAutoConfiguration {
-
-    @Bean
-    public Camunda7ProcessAdapter camunda7MessageAdapter() {
-        return new Camunda7ProcessAdapter();
-    }
-}
