+++ connect/connect-c7/process-adapter-in-c7/src/main/java/io/miragon/miranum/connect/adapter/in/c7/process/Camunda7ProcessAutoConfiguration.java
@@ -1,0 +1,14 @@
+package io.miragon.miranum.connect.adapter.in.c7.process;
+
+import org.camunda.community.rest.client.api.ProcessDefinitionApi;
+import org.springframework.boot.context.properties.EnableConfigurationProperties;
+import org.springframework.context.annotation.Bean;
+
+@EnableConfigurationProperties
+public class Camunda7ProcessAutoConfiguration {
+
+    @Bean
+    public Camunda7ProcessAdapter camunda7MessageAdapter(final ProcessDefinitionApi processDefinitionApi) {
+        return new Camunda7ProcessAdapter(processDefinitionApi);
+    }
+}
