+++ connect/connect-c7/process-adapter-in-c7/src/main/resources/META-INF/spring.factories
@@ -1,3 +1,3 @@
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-io.miragon.miranum.connect.adapter.in.c7.message.Camunda7ProcessAutoConfiguration
+io.miragon.miranum.connect.adapter.in.c7.process.Camunda7ProcessAutoConfiguration
