+++ examples/process-example/process-example-camunda-7/pom.xml
@@ -44,16 +44,6 @@
             <artifactId>camunda-bpm-spring-boot-starter-external-task-client</artifactId>
             <version>${camunda7.version}</version>
         </dependency>
-        <dependency>
-            <groupId>org.camunda.community</groupId>
-            <artifactId>camunda-engine-rest-client-openapi-springboot</artifactId>
-            <version>${camunda7.version}</version>
-        </dependency>
-<!--        <dependency>-->
-<!--            <groupId>org.camunda.community</groupId>-->
-<!--            <artifactId>camunda-engine-rest-client-openapi-springboot</artifactId>-->
-<!--            <version>${camunda7.version}</version>-->
-<!--        </dependency>-->
 
         <dependency>
             <groupId>org.springframework.boot</groupId>
