+++ examples/process-example/process-example-camunda-7/src/main/java/io/miragon/miranum/examples/process/c7/StartProcessController.java
@@ -1,40 +1,24 @@
 package io.miragon.miranum.examples.process.c7;
 
+import io.miragon.miranum.connect.process.application.port.in.StartProcessCommand;
+import io.miragon.miranum.connect.process.application.port.out.StartProcessPort;
 import lombok.AllArgsConstructor;
-import org.camunda.community.rest.client.api.ProcessDefinitionApi;
-import org.camunda.community.rest.client.dto.ProcessInstanceWithVariablesDto;
-import org.camunda.community.rest.client.dto.StartProcessInstanceDto;
-import org.camunda.community.rest.client.dto.VariableValueDto;
-import org.camunda.community.rest.client.invoker.ApiException;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-
-import java.util.HashMap;
-import java.util.Map;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequestMapping("api/process")
 @AllArgsConstructor
 public class StartProcessController {
 
-    private final ProcessDefinitionApi processDefinitionApi;
+    private final StartProcessPort startProcessPort;
 
     @PutMapping("/start")
-    public ResponseEntity<String> triggerProcessStart(@RequestBody StartProcessRequestDto startProcessRequestDto) {
-        try {
-
-            Map<String, VariableValueDto> variables = new HashMap<>();
-            variables.put("name", new VariableValueDto().value("name123").type("string"));
-            variables.put("key", new VariableValueDto().value("key123").type("string"));
-
-            ProcessInstanceWithVariablesDto processInstance = this.processDefinitionApi.
-                    startProcessInstanceByKey(startProcessRequestDto.getProcessKey(),
-                            new StartProcessInstanceDto().variables(variables));
-
-            return ResponseEntity.status(HttpStatus.OK).body("Started process with id " + processInstance.getId());
-        } catch (ApiException e) {
-            return ResponseEntity.internalServerError().build();
-        }
+    public ResponseEntity<Void> triggerProcessStart(@RequestBody StartProcessRequestDto startProcessRequestDto) {
+        startProcessPort.startProcess(new StartProcessCommand(startProcessRequestDto.getProcessKey(), startProcessRequestDto.getVariables()));
+        return ResponseEntity.ok().build();
     }
 }
